+++ CHANGES.rst
@@ -5,9 +5,13 @@ In next release ...
 
 Features:
 
+- Add support for configuring an RSS summary display of included
+  items.
+
 - Add support for Dexterity-based content.
 
 - Make content moderation requirement optional.
+
 
 1.2 (2012-03-09)
 ----------------